Testing Material Extended Icons updates
