+++ src/app/app.module.css
@@ -26,18 +26,6 @@
   flex-grow: 1;
   flex-basis: 0;
 }
-.group_2 {
-  justify-content: flex-start;
-  align-items: stretch;
-  align-content: flex-start;
-  gap: 16px;
-  position: relative;
-  padding: 16px;
-  min-width: 50px;
-  min-height: 50px;
-  flex-grow: 1;
-  flex-basis: 0;
-}
 .navbar {
   height: max-content;
   min-width: min-content;
@@ -47,6 +35,13 @@
   min-height: 0;
   max-width: 320px;
   flex-shrink: 0;
+}
+.view-container {
+  overflow: auto;
+  position: relative;
+  min-width: 0;
+  min-height: 0;
+  flex-grow: 1;
 }
 .nav-drawer::part(main) {
   width: 320px;
@@ -60,12 +55,4 @@
   font-size: 24px;
   width: 24px;
   height: 24px;
-}
-.button {
-  --ig-size: var(--ig-size-medium);
-  height: max-content;
-  flex-shrink: 0;
-}
-.button::part(base) {
-  background-color: var(--ig-secondary-100);
 }
