+++ src/app/app.test.tsx
@@ -1,9 +1,10 @@
 import { expect, test } from 'vitest';
+import { MemoryRouter } from 'react-router-dom';
 import { render } from '@testing-library/react';
 import App from './app';
 import 'element-internals-polyfill';
 
-test('renders without crashing', () => {
-  const wrapper = render(<App />);
+test('renders App component', () => {
+  const wrapper = render(<App />, { wrapper: MemoryRouter });
   expect(wrapper).toBeTruthy();
 });
