--- src/app/master-view/master-view.tsx
@@ -1,160 +1,0 @@
-import { IgrButton, IgrButtonModule, IgrIconButton, IgrIconButtonModule, IgrNavbar, IgrNavbarModule, IgrNavDrawer, IgrNavDrawerItem, IgrNavDrawerModule, IgrRipple, IgrRippleModule } from 'igniteui-react';
-import { useRef } from 'react';
-import styles from './master-view.module.css';
-import createClassTransformer from '../style-utils';
-
-IgrButtonModule.register();
-IgrIconButtonModule.register();
-IgrNavbarModule.register();
-IgrNavDrawerModule.register();
-IgrRippleModule.register();
-
-export default function MasterView() {
-  const classes = createClassTransformer(styles);
-  const uuid = () => crypto.randomUUID();
-  const navDrawer = useRef<IgrNavDrawer>(null);
-
-  return (
-    <>
-      <div className={classes("column-layout master-view-container")}>
-        <IgrNavbar className={classes("navbar")}>
-          <div style={{display: 'contents'}} slot="start" key={uuid()}>
-            <IgrIconButton variant="flat" clicked={() => navDrawer?.current?.toggle()}>
-              <span className={classes("material-icons")} key={uuid()}>
-                <span key={uuid()}>menu</span>
-              </span>
-              <IgrRipple key={uuid()}></IgrRipple>
-            </IgrIconButton>
-          </div>
-          <div className={classes("row-layout group")} key={uuid()}>
-            <h6 className={classes("h6")}>
-              <span>Testing Material Extended Icons</span>
-            </h6>
-          </div>
-          <div style={{display: 'contents'}} slot="end" key={uuid()}>
-            <IgrIconButton variant="flat">
-              <span className={classes("material-icons")} key={uuid()}>
-                <span key={uuid()}>search</span>
-              </span>
-              <IgrRipple key={uuid()}></IgrRipple>
-            </IgrIconButton>
-          </div>
-          <div style={{display: 'contents'}} slot="end" key={uuid()}>
-            <IgrIconButton variant="flat">
-              <span className={classes("material-icons")} key={uuid()}>
-                <span key={uuid()}>favorite</span>
-              </span>
-              <IgrRipple key={uuid()}></IgrRipple>
-            </IgrIconButton>
-          </div>
-          <div style={{display: 'contents'}} slot="end" key={uuid()}>
-            <IgrIconButton variant="flat">
-              <span className={classes("material-icons")} key={uuid()}>
-                <span key={uuid()}>more_vert</span>
-              </span>
-              <IgrRipple key={uuid()}></IgrRipple>
-            </IgrIconButton>
-          </div>
-        </IgrNavbar>
-        <div className={classes("row-layout group_1")}>
-          <IgrNavDrawer open="true" position="relative" ref={navDrawer} className={classes("nav-drawer")}>
-            <IgrNavDrawerItem key={uuid()}>
-              <span slot="icon" key={uuid()}>
-                <span className={classes("imx-icon imx-usa-party-democratic icon")} key={uuid()}></span>
-                <IgrRipple key={uuid()}></IgrRipple>
-              </span>
-              <div slot="content" key={uuid()}>Material Ext Icon</div>
-            </IgrNavDrawerItem>
-            <IgrNavDrawerItem key={uuid()}>
-              <span slot="icon" key={uuid()}>
-                <span className={classes("material-icons icon")} key={uuid()}>
-                  <span key={uuid()}>dashboard</span>
-                </span>
-                <IgrRipple key={uuid()}></IgrRipple>
-              </span>
-              <div slot="content" key={uuid()}>Nav Item 2</div>
-            </IgrNavDrawerItem>
-            <IgrNavDrawerItem key={uuid()}>
-              <span slot="icon" key={uuid()}>
-                <span className={classes("material-icons icon")} key={uuid()}>
-                  <span key={uuid()}>home</span>
-                </span>
-                <IgrRipple key={uuid()}></IgrRipple>
-              </span>
-              <div slot="content" key={uuid()}>Nav Item 3</div>
-            </IgrNavDrawerItem>
-          </IgrNavDrawer>
-          <div className={classes("row-layout group_2")}>
-            <span className={classes("imx-icon imx-coronavirus icon")}></span>
-            <IgrIconButton variant="flat">
-              <span className={classes("material-icons")} key={uuid()}>
-                <span key={uuid()}>ac_unit</span>
-              </span>
-              <IgrRipple key={uuid()}></IgrRipple>
-            </IgrIconButton>
-            <IgrButton type="button" className={classes("button")}>
-              <span className={classes("imx-icon imx-drop-shipping")} key={uuid()}></span>
-              <span key={uuid()}>Button</span>
-              <IgrRipple key={uuid()}></IgrRipple>
-            </IgrButton>
-          </div>
-          <IgrNavDrawer position="relative" className={classes("nav-drawer")}>
-            <div slot="mini" key={uuid()}>
-              <IgrNavDrawerItem key={uuid()}>
-                <span slot="icon" key={uuid()}>
-                  <span className={classes("material-icons icon")} key={uuid()}>
-                    <span key={uuid()}>account_circle</span>
-                  </span>
-                  <IgrRipple key={uuid()}></IgrRipple>
-                </span>
-              </IgrNavDrawerItem>
-              <IgrNavDrawerItem key={uuid()}>
-                <span slot="icon" key={uuid()}>
-                  <span className={classes("material-icons icon")} key={uuid()}>
-                    <span key={uuid()}>dashboard</span>
-                  </span>
-                  <IgrRipple key={uuid()}></IgrRipple>
-                </span>
-              </IgrNavDrawerItem>
-              <IgrNavDrawerItem key={uuid()}>
-                <span slot="icon" key={uuid()}>
-                  <span className={classes("material-icons icon")} key={uuid()}>
-                    <span key={uuid()}>home</span>
-                  </span>
-                  <IgrRipple key={uuid()}></IgrRipple>
-                </span>
-              </IgrNavDrawerItem>
-            </div>
-            <IgrNavDrawerItem key={uuid()}>
-              <span slot="icon" key={uuid()}>
-                <span className={classes("material-icons icon")} key={uuid()}>
-                  <span key={uuid()}>account_circle</span>
-                </span>
-                <IgrRipple key={uuid()}></IgrRipple>
-              </span>
-              <div slot="content" key={uuid()}>Nav Item 1</div>
-            </IgrNavDrawerItem>
-            <IgrNavDrawerItem key={uuid()}>
-              <span slot="icon" key={uuid()}>
-                <span className={classes("material-icons icon")} key={uuid()}>
-                  <span key={uuid()}>dashboard</span>
-                </span>
-                <IgrRipple key={uuid()}></IgrRipple>
-              </span>
-              <div slot="content" key={uuid()}>Nav Item 2</div>
-            </IgrNavDrawerItem>
-            <IgrNavDrawerItem key={uuid()}>
-              <span slot="icon" key={uuid()}>
-                <span className={classes("material-icons icon")} key={uuid()}>
-                  <span key={uuid()}>home</span>
-                </span>
-                <IgrRipple key={uuid()}></IgrRipple>
-              </span>
-              <div slot="content" key={uuid()}>Nav Item 3</div>
-            </IgrNavDrawerItem>
-          </IgrNavDrawer>
-        </div>
-      </div>
-    </>
-  );
-}
