+++ src/app/page2/page2.test.tsx
@@ -1,0 +1,9 @@
+import { expect, test } from 'vitest';
+import { render } from '@testing-library/react';
+import Page2 from './page2';
+import 'element-internals-polyfill';
+
+test('renders Page2 component', () => {
+  const wrapper = render(<Page2 />);
+  expect(wrapper).toBeTruthy();
+});
